% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwaySlim-package.r
\docType{package}
\name{pathwaySlim}
\alias{pathwaySlim}
\alias{pathwaySlim-package}
\title{pathwaySlim.
Provides slim pathway to gene annotations for mouse and human}
\description{
pathwaySlim.
Provides slim pathway to gene annotations for mouse and human
}

